+++ src/components/App/modals/commands.tsx
@@ -4,20 +4,19 @@ import { Box, Divider, Group, Modal, ScrollArea, Stack, Text, TextInput } from "
 import { useInputState } from "@mantine/hooks";
 import clsx from "clsx";
 import posthog from "posthog-js";
-import { type KeyboardEvent, useMemo, useRef, useState } from "react";
+import { type KeyboardEvent, useMemo, useRef } from "react";
 import { adapter } from "~/adapter";
 import { Entry } from "~/components/Entry";
 import { Icon } from "~/components/Icon";
 import { PreferenceInput } from "~/components/Inputs/preference";
-import { Shortcut } from "~/components/Shortcut";
 import { Spacer } from "~/components/Spacer";
 import { useBoolean } from "~/hooks/boolean";
 import { useKeymap } from "~/hooks/keymap";
 import { useKeyNavigation } from "~/hooks/keys";
 import { useStable } from "~/hooks/stable";
 import { dispatchIntent, useIntent } from "~/hooks/url";
+import { type Command, useCommandCategories } from "~/providers/Commands";
 import { useConfigStore } from "~/stores/config";
-import { type Command, type CommandCategory, computeCommands } from "~/util/commands";
 import { ON_STOP_PROPAGATION, Y_SLIDE_TRANSITION, fuzzyMatch } from "~/util/helpers";
 import { iconOpen, iconSearch } from "~/util/icons";
 
@@ -27,7 +26,8 @@ export function CommandPaletteModal() {
 
 	const [isOpen, openHandle] = useBoolean();
 	const [search, setSearch] = useInputState("");
-	const [categories, setCategories] = useState<CommandCategory[]>([]);
+
+	const categories = useCommandCategories();
 
 	const handlePreferenceInput = useStable((e: KeyboardEvent) => {
 		e.stopPropagation();
@@ -123,9 +123,7 @@ export function CommandPaletteModal() {
 
 	useIntent("open-command-palette", () => {
 		openHandle.open();
-
 		setSearch("");
-		setCategories(computeCommands());
 	});
 
 	useKeymap([
@@ -217,7 +215,8 @@ export function CommandPaletteModal() {
 													ml={-8}
 												/>
 											)}
-											{cmd.shortcut && (
+											{/* TODO Use keybindings */}
+											{/* {cmd.binding && (
 												<>
 													<Spacer />
 													<Group gap="lg">
@@ -232,7 +231,7 @@ export function CommandPaletteModal() {
 														))}
 													</Group>
 												</>
-											)}
+											)} */}
 											{cmd.action.type === "preference" && (
 												<>
 													<Spacer />
